Move the test to its own file so it is not masked by an unrelated pre-existing failure in node-http.test.ts
