+++ test/js/node/http/node-http.test.ts
@@ -4054,47 +4054,3 @@ it("OutgoingMessage outputData is per-instance and _flushOutput is defined", () 
   c.outputData.push({ data: "y", encoding: "utf8", callback: null });
   expect(d.outputData.length).toBe(0);
 });
-
-it("aborted request body emits 'error' ECONNRESET and res 'close' before req 'close'", async () => {
-  // Node.js's socketOnClose → abortIncoming destroys the in-flight request with
-  // a ConnResetException, so req emits 'error' (when listened for) and 'close'
-  // after res 'close'. A mid-body disconnect is surfaced as a clientError which
-  // destroys the socket before the native close; the request abort must still
-  // go through the socket-close path, not a no-error destroy earlier.
-  const events: string[] = [];
-  const { promise: gotRequest, resolve: resolveRequest } = Promise.withResolvers<void>();
-  const { promise: reqClosed, resolve: resolveReqClosed } = Promise.withResolvers<void>();
-  const { promise: resClosed, resolve: resolveResClosed } = Promise.withResolvers<void>();
-
-  const server = createServer((req, res) => {
-    req.on("aborted", () => events.push("req.aborted"));
-    req.on("error", e => events.push("req.error:" + (e as NodeJS.ErrnoException).code));
-    req.on("close", () => {
-      events.push("req.close");
-      resolveReqClosed();
-    });
-    res.on("close", () => {
-      events.push("res.close");
-      resolveResClosed();
-    });
-    req.on("data", () => {});
-    resolveRequest();
-  });
-  try {
-    server.listen(0, "127.0.0.1");
-    await once(server, "listening");
-    const { port } = server.address() as AddressInfo;
-
-    const client = connect(port, "127.0.0.1");
-    client.on("error", () => {});
-    await once(client, "connect");
-    client.write("POST / HTTP/1.1\r\nHost: x\r\nContent-Length: 100\r\n\r\npartial");
-    await gotRequest;
-    client.destroy();
-    await Promise.all([reqClosed, resClosed]);
-
-    expect(events).toEqual(["req.aborted", "res.close", "req.error:ECONNRESET", "req.close"]);
-  } finally {
-    server.close();
-  }
-});
